+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 ## 3.2.0
 - Run a ping from the list (new settings option)
+- Update libraries
+- Add more user device info (Build.DEVICE, Build.ID, Build.BOARD)